move server/needles to server/support/needles and convert to typescript
