+++ generators/server/index.mts
@@ -22,7 +22,6 @@ import { SpringBootApplication } from './types.mjs';
 
 export { default } from './generator.mjs';
 export { serverFiles as files } from './files.mjs';
-export * from './needles.mjs';
 
 // TODO move to ./generator.mts
 export type ApplicationDefinition = {
